change directory structure
aws-sdk-go and strings utility all in a directory. It's difficult for me to manage them.
So I change directory structure.
