+++ pkg/awsapi/ec2_instance_connect.go
@@ -1,4 +1,4 @@
-package omssh
+package awsapi
 
 import (
 	"log"
